+++ src/Text/Regex/Do/Trim.hs
@@ -5,6 +5,8 @@ import Data.Char(isSpace)
 import qualified Data.ByteString as B
 import Text.Regex.Do.Convert
 import Text.Regex.Do.Pcre.Replace
+import Text.Regex.Do.Type.MatchHint
+
 
 {- | removes leading and trailing spaces and tabs   -}
 
@@ -13,7 +15,7 @@ class Trim a where
 
 
 instance Trim B.ByteString where
-    trim bs1 = replace [All] rx1 repl1 $ Body bs1
+    trim bs1 = replace (All []) rx1 repl1 $ Body bs1
        where repl1 = Replacement B.empty
              rx1 = rxFn "(^[\\s\\t]+)|([\\s\\t]+$)"
              rxFn = Pattern . toByteString